prevent previous startpage deletion when keep_startpage is checked, append links to previous page if necessary, avoid duplicates

+++ core/generate.php
@@ -49,13 +49,59 @@ function generate_error(){
     echo "Generating Error file<br><br>";
 }
 
-function generate_start($start_page){
+function generate_start($start_page, $keep_startpage, $append_links){
     global $startfile;
-    $step0 = str_replace("{TABLE_BUTTONS}", $start_page, $startfile);
-    $destination_file = fopen("app/index.php", "w") or die("Unable to open file!");
-    fwrite($destination_file, $step0);
-    fclose($destination_file);
-    echo "Generating Startpage file<br>";
+    global $generate_start_checked_links;
+    global $startpage_filename;
+
+    if (!$keep_startpage) {
+        $step0 = str_replace("{TABLE_BUTTONS}", $start_page, $startfile);
+        $destination_file = fopen($startpage_filename, "w") or die("Unable to open fresh startpage file!");
+        fwrite($destination_file, $step0);
+        fclose($destination_file);
+        echo "Generating Startpage file<br>";
+    } else {
+        if ($append_links) {
+            // load existing template
+            echo "Retrieving existing Startpage file<br>";
+            $handle = fopen($startpage_filename, "r") or die("Unable to open existing startpage file!");;
+            $startfile = fread($handle, filesize($startpage_filename));
+            fclose($handle);
+            
+            // extract existing links from app/index.php
+            echo "Looking for new links to append to Startpage file<br>";
+            $link_matcher_pattern = '/href=["\']?([^"\'>]+)["\']?/im';
+            preg_match_all($link_matcher_pattern, $startfile, $startfile_links);
+            if (count($startfile_links)) {
+                foreach($startfile_links[1] as $startfile_link) {
+                    // echo '- Found existing link '.$startfile_link.'<br>';
+                }
+            }
+
+            // do not append links to app/index.php if they  already 
+            preg_match_all($link_matcher_pattern, $start_page, $start_page_links);
+            if (count($start_page_links)) {
+                foreach($start_page_links[1] as $start_page_link) {
+                    if (!in_array($start_page_link, $generate_start_checked_links)) {
+                        if (in_array($start_page_link, $startfile_links[1])) {
+                            echo '- Not appending '.$start_page_link.' as it already exists<br>';
+                        } else {
+                            echo '- Appending '.$start_page_link.'<br>';
+                            array_push($startfile_links[1], $start_page_link);
+                            $linkname = str_replace('-index.php', '', basename($start_page_link));
+                            $step0 = preg_replace('/<\/div>.*<\/center>/msx', "\t".'<a href="'.$start_page_link.'>" class="btn btn-primary" role="button">'.$linkname.'</a>'."\n</div>\n</center>", $startfile);
+                            $destination_file = fopen($startpage_filename, "w") or die("Unable to open file!");
+                            fwrite($destination_file, $step0);
+                            fclose($destination_file);
+                        }
+                        array_push($generate_start_checked_links, $start_page_link);
+                    }
+                }
+            }
+        }
+    }
+
+    
 }
 
 function generate_index($tablename,$tabledisplay,$index_table_headers,$index_table_rows,$column_id, $columns_available, $index_sql_search) {
